make quitting reasonable
