+++ demo/game/start-web-client.js
@@ -1,0 +1,12 @@
+const { generateClientServer } = require("../../src/generate-client-server.js");
+const ClientClass = require("./src/core/game-client.js");
+const ServerClass = require("./src/core/game-server.js");
+const ClientServer = generateClientServer(ClientClass, ServerClass);
+
+const url = `http://localhost:8080`;
+const public = `${__dirname}/public`;
+const webclient = ClientServer.createWebClient(url, public);
+webclient.listen(0, () => {
+  console.log(`web client listening on ${webclient.address().port}`);
+});
+
